Add clear todo data feature and improve Json handling
This commit introduces a new feature in the Todo application to clear all existing Todo data. The main menu now includes the option to clear out all data from the Todo list. This is reflected in the changes made in MainMenu.cs and InputHandler.cs.

The naming of the function responsible for toggling the completion status of Todos has been updated to be more intuitive and meaningful. This change can be seen in TodoManager.cs and InputManager.cs.

Lastly, the handling, validation, and loading of Json data have been improved to provide better user experience and improve application reliability. Changes to facilitate this can be seen in JsonHandler.cs and Program.cs. Validation of Json data now includes checking for an empty Json object and provides appropriate user feedback.