.Add feature to update Todo priority
Implemented feature to update priority of a Todo within TodoApp. This includes creation of new prompts for user to input desired priority, display the changed priority to the user, handling the priority change request in the InputHandler, updating the priority within TodoManager and adjusting the MainMenu prompt. This feature provides user with an additional functionality to modify the priority of existing Todos as per their requirement.